requirements.txt
websocket-client
msgpack
